+++ Changelog.md
@@ -20,6 +20,7 @@
   - Advanced layout &mdash; Uses simulated annealing to find better layouts
 - Slice preview view direction selector (lcd screen vs build plate)
 - Support for Spacenav input (3Dconnexion 3D input device)
+- Initial camera distance increased
 
 ## v0.7.0 &mdash; May 7th, 2026